Try to appease the go.work and replace gods
For some reason make go-mod-tidy is trying to find proto-public v1alpha1
package, despite it not being required by the failing module
(consul-container). My guess is that this is due to the latest tag
containing v1alpha1 code and the replace is not playing nicely.

I'm guessing the require directive for v0.5.1 of proto-public is causing
this, and an upgrade to the latest commit will fix it.